requirements.txt
requests==2.31.0
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==8.0.1
